SmoothedSpectralAbscissa.jl

The Purpose of this package is computing the Smoothed Spectral Abscissa of square matrices, and the associated gradient. The computation can be optimized for recursion, so that each new calculation does not reallocate memory.

I implemented only the "simplified" version that does not have projections.

The algorithm is described in the following paper:

The Smoothed Spectral Abscissa for Robust Stability Optimization , J Vanbiervliet et al , 2009. DOI: 10.1137/070704034

TO-DOs

Add references

describe the algorithm

SmoothedSpectralAbscissa.SSAAlloc

SmoothedSpectralAbscissa.PQ_init!

SmoothedSpectralAbscissa.default_eps_ssa

SmoothedSpectralAbscissa.ssa_simple

SmoothedSpectralAbscissa.ssa_simple!

SmoothedSpectralAbscissa.ssa_simple_newton

SmoothedSpectralAbscissa.ssa_simple_withgradient

SmoothedSpectralAbscissa.SSAAlloc — Method

    SSAAlloc{M<:AbstractMatrix,V<:AbstractVector}(n::Integer) -> SSAAlloc

Allocation structure that makes SSA computation mode efficient. n is the size of the matrix (or matrices) on which one wants to compute the SSA

SmoothedSpectralAbscissa.PQ_init! — Method

    PQ_init!(A::AbstractMatrix,PQ::SSAAlloc) -> nothing

Prepares SSAAlloc with the matrix on which one wants to compute the SSA. Schur decompositions are performed here.

SmoothedSpectralAbscissa.default_eps_ssa — Method

    default_eps_ssa(A::AbstractMatrix{Float64}) -> eps_ssa

default value for the $\varepsilon$ used to compute the SSA

SmoothedSpectralAbscissa.ssa_simple — Function

    ssa_simple(A::AbstractMatrix , ssa_eps::Union{Nothing,Float64}=nothing)
        -> ssa::Float64

Computes the Smoothed Spectral Abscissa (SSA) of matrix A in its simplest formulation If ssa_eps is not specified, a default value is used.

In case you need to call the SSA iteratively, consider the version that acts on pre-allocated memory.

SmoothedSpectralAbscissa.ssa_simple! — Function

    ssa_simple!(A::AbstractMatrix, with_gradient::Bool,PQ::SSAAlloc ,
        ssa_eps::Union{Nothing,Float64}=nothing) -> ssa::Float64

Computes the Smoothed Spectral Abscissa (SSA) of matrix A in its simplest formulation. For efficiency, A is reallocated in the computation. If ssa_eps is not specified, a default value is used.

SmoothedSpectralAbscissa.ssa_simple_newton — Function

This uses the Newton method and the analytic derivative , seems as fast as the Order16 method

SmoothedSpectralAbscissa.ssa_simple_withgradient — Function

  ssa_simple_withgradient(A::AbstractMatrix , ssa_eps::Union{Nothing,Float64}=nothing)
        -> (ssa,gradient_matrix)::Tuple{Float64,Matrix}

Computes the Smoothed Spectral Abscissa (SSA) of matrix A in its simplest formulation, and also its gradient, for each matrix element. If ssa_eps is not specified, a default value is used.

In case you need to call the SSA iteratively (e.g. gradient-based optimization), consider the version that acts on pre-allocated memory.